Add trailing NULL's to the varargs functions as required.
(John Spencer)

    readline.c r1.106

Bounds search for reallocated index, from OpenBSD via Andreas Fett

    map.c 1.34
    map.h 1.10
    parse.c 1.27
    read.c 1.71

This is a direct commit to stable/10 as these changes are included in a
broader sync with NetBSD committed by bapt@ in r276881.

Reviewed by:	bapt
Differential Revision:	https://reviews.freebsd.org/D1728

+++ lib/libedit/map.h
@@ -57,7 +57,7 @@ typedef struct el_map_t {
 	int		 type;		/* Emacs or vi			*/
 	el_bindings_t	*help;		/* The help for the editor functions */
 	el_func_t	*func;		/* List of available functions	*/
-	int		 nfunc;		/* The number of functions/help items */
+	size_t		 nfunc;		/* The number of functions/help items */
 } el_map_t;
 
 #define	MAP_EMACS	0
